+++ client/js/node.js
@@ -19,8 +19,16 @@ const processBalance = async () => {
 
     if (status && status.data && status.data.account && status.data.account.balance) {
         const {total, liquid} = status.data.account.balance
+
+        globalThis.balance = +total / 10**9
+
         $("#balance-total").text((total/10**9).format(2, null, ",", "."))
         $("#balance-liquid").text((liquid/10**9).format(2, null, ",", "."))
+
+        console.log(globalThis.balance, globalThis.price)
+
+        $("#balance-usd").text((globalThis.balance * globalThis.price).format(2, null, ",", "."))
+
     }
 }