+++ server/monitor.mjs
@@ -12,6 +12,7 @@ import {processAlerter} from "./alerter.mjs"
 import {processBalanceSend} from "./balance-sender.mjs"
 import {processHello} from "./hello.mjs"
 import {getUptime} from "./uptime.mjs"
+import {getLedgerInfo} from "./ledger.mjs";
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 const configPath = path.resolve(__dirname, 'config.json')
@@ -50,7 +51,7 @@ const requestListener = async (req, res) => {
         case '/blockchain': response = await nodeInfo('blockchain', config); break;
         case '/node-status': response = await nodeInfo('node-status', config); break;
         case '/balance': response = await nodeInfo('balance', config); break;
-        case '/delegators': response = await nodeInfo('delegators', config); break;
+        case '/delegators': response = await getLedgerInfo("delegators", config); break;
         case '/block-speed': response = await nodeInfo('block-speed', config); break;
         case '/explorer': response = await getExplorerSummary(); break;
         case '/uptime': response = await getUptime(config.publicKey); break;